+++ core/src/routes/(editor)/editor/+page.svelte
@@ -1,9 +1,13 @@
 <script>
     import "../../../app.css";
     import { onMount } from "svelte";
-    import { globalEditorPreferencesStore, globalComponentCollectionStore } from "../../globals/globalstores.js";
+    import { globalEditorPreferencesStore, globalComponentCollectionStore, globalThemeStore } from "../../globals/globalstores.js";
+    import { globalSelectedElementStore } from "../../globals/selectorstores.js";
+    import { getComponent } from "../../globals/globalfunctions.js";
+    import { sendDeletedElement } from "../../(shared)/shared/sharedfunctions.js";
     import Editortree from "./editortree.svelte";
     import Selector from "../../(shared)/shared/selector.svelte";
+    import swal from 'sweetalert';
 
     /// ATTENTION ! 
 
@@ -16,6 +20,37 @@
     let globalComponentCollection = $globalComponentCollectionStore;
     $: globalComponentCollectionStore.set(globalComponentCollection);
 
+    let globalTheme = $globalThemeStore;
+    $: globalThemeStore.set(globalTheme);
+
+      
+    // TODO: import all modules dynamically
+
+    import Body from "../../(modules)/modules/body.svelte";
+    import Div from "../../(modules)/modules/div.svelte";
+    import Text from "../../(modules)/modules/text.svelte";
+
+    /**
+     * Definition and list of all modules in a JSON.
+     * @returns {JSON}
+     * 
+     */
+    const JsonOfModules = {
+        "body": Body,
+        "div": Div,
+        "text": Text,
+    };
+
+    /**
+     * Definition and list of all modules in a JSON.
+     * @returns {JSON}
+     * 
+     */
+     const JsonOfTypes = {
+        "body": {"title": "Body", "data": ""},
+        "div": {"title": "Container", "data": ""},
+        "text": {"title": "Text", "data": "text"},
+    };
 
     onMount(() => {
 
@@ -34,35 +69,90 @@
                 /// Set data as global writables
                 globalComponentCollection = data.componentCollection;
                 globalEditorPreferences = data.editorPreferences;
+                globalTheme = data.globalTheme;
             }
+        });
+
+        window.addEventListener("keydown", (e) => {
+            if(e.key === 'Delete' && $globalSelectedElementStore !== null){
+
+                const comp = getComponent($globalComponentCollectionStore, $globalSelectedElementStore.id);
+                // var question = "Delete \"" + JsonOfTypes[comp.type].title + "\" ?\n\n";
+                // question += JsonOfTypes[comp.type].data != "" ? comp.data[JsonOfTypes[comp.type].data] : "";
+
+                // if(confirm(question)) {
+
+                // }
+                if(comp.type != "body"){
+                    swal({
+                        title: "Delete \"" + JsonOfTypes[comp.type].title + "\" ?",
+                        text: JsonOfTypes[comp.type].data != "" ? (comp.data[JsonOfTypes[comp.type].data] ?? "") : "",
+                        icon: "warning",
+                        buttons: ["No, keep it", "Yes, DELETE"],
+                        dangerMode: true,
+                    })
+                    .then((willDelete) => {
+                        if (willDelete) {
+                            sendDeletedElement(comp.uuid);
+                            swal("Element deleted", {
+                                icon: "success",
+                            });
+                        } else {
+                            swal("Element kept!");
+                        }
+                    });
+                }
+
+            }
+        });
+
+        window.addEventListener("error", (e) => {
+
+            swal({
+                title: "Error on Editor Panel",
+                text: e.error.toString(),
+                icon: "error",
+            });
+
         });
 
     });
 
-   
-    // TODO: import all modules dynamically
+    // import { writable } from "svelte/store";
+    // let overlayBackground = writable("#242425");
+    // let modalBackground = writable("#242425");
+    // let modalBorder = writable("1px solid #242425");
+    // let modalCustomCss = writable("");
 
-    import Body from "../../(modules)/modules/body.svelte";
-    import Div from "../../(modules)/modules/div.svelte";
-    import Text from "../../(modules)/modules/text.svelte";
+    // $: $globalThemeStore, (() => {
+    //     console.log("$globalThemeStore updated!");
+    //     if($globalThemeStore !== undefined){
+    //         console.log("$globalThemeStore !== undefined");
+    //         overlayBackground.set($globalThemeStore.swal.overlay.backgroundColor != undefined ? $globalThemeStore.swal.overlay.backgroundColor : "#242425");
+    //         modalBackground.set($globalThemeStore.swal.modal.backgroundColor != undefined ? $globalThemeStore.swal.modal.backgroundColor : "#242425");
+    //         modalBorder.set($globalThemeStore.swal.modal.border != undefined ? $globalThemeStore.swal.modal.border : "1px solid #242425");
+    //         modalCustomCss.set($globalThemeStore.swal.modal.customCss != undefined ? $globalThemeStore.swal.modal.customCss : "");
 
-    /**
-     * Definition and list of all modules in a JSON.
-     * @returns {JSON}
-     * 
-     */
-    const JsonOfModules = {
-        "body": Body,
-        "div": Div,
-        "text": Text,
-    };
+    //     }
+    //     console.log('--swalOverlayBackgroundColor:'+$overlayBackground+';');
+    //     console.log('--swalModalBackgroundColor:'+$modalBackground+';');
+    //     console.log('--swalModalBorder:'+$modalBorder+';');
+    //     console.log('--swalModalCustomCss:'+$modalCustomCss+';');
+    // })();
 
 
 </script>
 
+<!-- style='
+    --swalOverlayBackgroundColor:{$overlayBackground};
+   --swalModalBackgroundColor:{$modalBackground};
+   --swalModalBorder:{$modalBorder};
+   --swalModalCustomCss:{$modalCustomCss};
+' -->
 
+<div id="editorInnerPanel" 
 
-<div id="editorInnerPanel">
+>
 
     {#if $globalComponentCollectionStore}
     {#each $globalComponentCollectionStore as component}
@@ -75,8 +165,10 @@
     {/if}
 
 
-
+fatih
 </div> <!-- editorInnerPanel -->
+
+<!-- <div class="hidden swal-overlay swal-modal"></div> -->
     
 <svelte:component this={Selector} />
 
@@ -86,5 +178,16 @@
         width: 100%; 
         height: 100vh; /* subject to change in following versions */
     }
+
+        
+    /* :global(.swal-overlay) {
+        background-color: var(--swalOverlayBackgroundColor) !important; 
+    }
+    
+    :global(.swal-modal) {
+        background-color: var(--swalModalBackgroundColor) !important;
+        border: var(--swalModalBorder) !important;
+    } */
+
 
 </style>